tsm/pyproject.toml
[tool.poetry]
name = "tsm"
version = "0.1.0"
description = ""
authors = ["Dan Zilberman <dan.zilberman99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.9.0"}
docker = "^7.0.0"
pyyaml = "^6.0.1"
pydantic = "^2.5.2"

[tool.poetry.scripts]
tsm = 'tsm.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
